Introduce GPL2.0 and MIT, package rename
