--- aurae-ebpf/rust-toolchain.toml
@@ -1,6 +1,0 @@
-[toolchain]
-channel="nightly"
-targets = [ "x86_64-unknown-linux-musl" ]
-# The source code of rustc, provided by the rust-src component, is needed for
-# building eBPF programs.
-components = [ "rustc", "rust-std", "cargo", "rust-docs", "rustfmt", "clippy", "rust-src" ]
