+++ auraed/Cargo.toml
@@ -43,7 +43,7 @@ path = "src/bin/main.rs"
 anyhow = { workspace = true }
 aurae-client = { workspace = true }
 aurae-proto = { workspace = true }
-aurae-ebpf-shared = { path = "../aurae-ebpf-shared" }
+aurae-ebpf-shared = { path = "../ebpf-shared" }
 aya = { version = ">=0.11", features=["async_tokio"] }
 aya-log = "0.1"
 backoff = { version = "0.4.0", features = ["tokio"] }
